max77665-flash: Fix print formatting

+++ drivers/media/platform/tegra/max77665-flash.c
@@ -1345,8 +1345,8 @@ static int max77665_f_regulator_get(struct max77665_f_info *info,
 
 	reg = regulator_get(info->dev, vreg_name);
 	if (unlikely(IS_ERR(reg))) {
-		dev_err(info->dev, "%s %s ERR: %d\n",
-			__func__, vreg_name, (int)reg);
+		dev_err(info->dev, "%s %s ERR: %zd\n",
+			__func__, vreg_name, (intptr_t)reg);
 		err = PTR_ERR(reg);
 		reg = NULL;
 	} else
@@ -1425,7 +1425,7 @@ static void max77665_f_caps_layout(struct max77665_f_info *info)
 	info->torch_timeouts[1] = start_ptr + max77665_f_torch_cap_size;
 
 	info->torch_cap_size = MAX77665_TORCH_CAP_TIMEOUT_SIZE;
-	dev_dbg(info->dev, "%s: %d(%d + %d), %d(%d + %d)\n", __func__,
+	dev_dbg(info->dev, "%s: %d(%zu + %zu), %d(%zu + %zu)\n", __func__,
 		info->flash_cap_size, max77665_f_flash_cap_size,
 		max77665_f_flash_timeout_size, info->torch_cap_size,
 		max77665_f_torch_cap_size, max77665_f_torch_timeout_size);
